Start fsp subpackage, separate momo stuff
